requirements.txt
PyYAML~=6.0
diot~=0.1.6
pandas~=2.0.3
requests~=2.28.1
beautifulsoup4~=4.8.0
tqdm~=4.64.1
pytest~=7.1.3
setuptools~=61.3.1
python-dateutil~=2.8.2
numpy==1.26.4
lxml
python-telegram-bot<=18.0.0